README.md: Working toward basic documentation (#10)
* README.md: document config file

Still pretty sparse, but as this bit is necessary to actually use
pi-gen, here it is.

* README.md: Overview of Raspbian's stages

* README.md: Stages -> Raspbian Stages + heading

Decided that the Stages heading probably made more sense as being
retitled to refer specifically to Raspbian since it has no direct
bearing on pi-gen itself which could produce any Debianish system for
the pi, and also that there ought to be a Stage Anatomy section first
talking about the nuts and bolts of what goes into these stages.

Stage Anatomy is just a heading for now.

* README.md: Make Config section less conversational

Left the stage overview as it was since it's intended more as a plain
language explanation than a technically complete look.  I believe that
kind of information is useful to have, but perhaps as README.md becomes
longer it may be advisable to move it to another location.